+++ src/components/common/CouldNotLoad/README.md
@@ -1,0 +1,6 @@
+# Could Not Load
+
+`CouldNotLoad` takes in two inputs:
+
+- an error message such as `You are currently not in a section.`
+- a className in order to style freely based off the error. We may want some errors styled differently based off their container page.
