+++ src/components/common/CouldNotLoad/index.ts
@@ -1,0 +1,1 @@
+export { default as CouldNotLoad } from './CouldNotLoad';
