--- src/components/pages/README.md
@@ -1,4 +1,0 @@
-# Could Not Load
-
-A component that will render if a Teacher or Student are not in a section yet.
-Most use cases would be non-clever users, which we may find a different solution for in the near future.
