+++ src/components/pages/StudentDashboard/StudentSectionList/StudentSectionListContainer.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { Sections } from '../../../../api';
 import { auth, sections } from '../../../../state';
-import CouldNotLoad from '../../CouldNotLoad';
+import CouldNotLoad from '../../../common/CouldNotLoad/CouldNotLoad';
 import RenderStudentSectionList from './RenderStudentSectionList';
 
 const StudentSectionListContainer = (): React.ReactElement => {
